requirements.txt
fastapi
uvicorn==0.24.0
google-api-python-client==2.110.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
python-multipart==0.0.6
httpx>=0.27.1
nest-asyncio==1.6.0
mcp>=1.22.0
